+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 ### gem internals
  - Finder Test Helper change from `assert_no_offense` to `assert_no_offenses`
  - auto-inclusion of a gem's helper method class to Finders under a gem's namespace
- 
+ - support gemspec parsing
+ - better file finding and error handling
+
 ## [0.1.3] - 2026-08-24
  - Added Finder for: "Passing `binds` as a positional argument to `insert` deprecated", see [original commit](https://github.com/rails/rails/commit/2dca5457ab2097626481fdbec233ea56d3fb9ee3)
  - Added `effort` to the CLI output